PF4 Masthead Work (#377)
* fix layout issues and update logo

* fix failing tests

* fix dashboard layout after masthead changeover
