+++ src/pages/landing/landingPageMastHead.js
@@ -1,15 +1,13 @@
 import * as React from 'react';
 
 const LandingPageMastHead = () => (
-  <header className="row integr8ly-landing-page-masthead pf-c-content">
-    <div className="col-xs-12">
-      <h1 className="pf-u-mt-lg">Welcome to the Red Hat Solution Explorer</h1>
-      <p>
-        Get started with an end-to-end solution walkthrough or
-        <br />
-        use any of the available application services to create custom integrations.
-      </p>
-    </div>
+  <header className="integr8ly-landing-page-masthead pf-c-content">
+    <h1 className="pf-u-mt-lg">Welcome to the Red Hat Solution Explorer</h1>
+    <p>
+      Get started with an end-to-end solution walkthrough or
+      <br />
+      use any of the available application services to create custom integrations.
+    </p>
   </header>
 );
