+++ src/router/router.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Redirect, Route, Switch } from 'react-router-dom';
+import { BackgroundImage, BackgroundImageSrc } from '@patternfly/react-core';
 import { routes } from '../routes';
 import { buildProvisioningScreen } from '../components/provisioning/provisioning';
 
@@ -43,8 +44,18 @@ class Router extends React.Component {
   render() {
     const { renderRoutes, redirectRoot } = Router.renderRoutes();
 
+    const bgImages = {
+      [BackgroundImageSrc.lg]: '/assets/images/pfbg_1200.jpg',
+      [BackgroundImageSrc.sm]: '/assets/images/pfbg_768.jpg',
+      [BackgroundImageSrc.sm2x]: '/assets/images/pfbg_768@2x.jpg',
+      [BackgroundImageSrc.xs]: '/assets/images/pfbg_576.jpg',
+      [BackgroundImageSrc.xs2x]: '/assets/images/pfbg_576@2x.jpg',
+      [BackgroundImageSrc.filter]: '/assets/images/background-filter.svg#image_overlay'
+    };
+
     return (
       <div className="integr8ly-container">
+        <BackgroundImage src={bgImages} />
         <Switch>
           {renderRoutes}
           {redirectRoot}
