+++ src/styles/application/components/_landingPage.scss
@@ -1,7 +1,6 @@
 .integr8ly-landing-page-tutorial-dashboard-section {
   display: flex;
   flex-wrap: wrap;
-  height: calc(100vh - 170px); // View port height - mastheads
 
   &-left {
     flex-basis: 66.6%;
