+++ src/styles/application/components/_sidePanel.scss
@@ -10,36 +10,3 @@
     list-style: none;
   }
 }
-
-// .integr8ly-helpful {
-//   &-links {
-//     position: fixed;
-//     top: 0;
-//     right: 0;
-//     bottom: 0;
-//     background-color: $pf-color-black-150;
-//     border-left: var(--pf-global--BorderWidth--sm) solid;
-//     border-left-color: $pf-color-black-300;
-
-//     &-product-title {
-//       i {
-//         font-size: var(--pf-global--FontSize--md);
-
-//         /* stylelint-disable */
-//         /* only 3 levels of nesting allowed - this needs to be fixed */
-//         &.pficon-on-running {
-//           color: $pf-color-green-400;
-//         }
-
-//         &.fa-pie-chart {
-//           color: $pf-color-black-600;
-//         }
-
-//         &.pficon-pending {
-//           color: $pf-color-orange-400;
-//         }
-//         /* stylelint-enable */
-//       }
-//     }
-//   }
-// }
