+++ src/styles/application/components/_tutorialCard.scss
@@ -23,6 +23,8 @@
     position: absolute;
     width: 100%;
     bottom: 0;
+    padding-right: 10px;
+    padding-left: 10px;
     padding-bottom: 20px;
 
     .progress { /* stylelint-disable-line selector-class-pattern */
