+++ src/styles/application/components/_tutorialDashboard.scss
@@ -1,7 +1,5 @@
 .integr8ly-tutorial-dashboard {
-  height: 100vh;
-  margin-bottom: 0;
-  background-color: $pf-color-black-150;
+  background-color: $pf-color-black-200;
   border: 0;
   box-shadow: none;
 
@@ -21,6 +19,13 @@
     }
   }
 
+  /* stylelint-disable */
+  & .panel-heading,
+  .cards-pf {
+    background-color: $pf-color-black-200;
+  }
+  /* stylelint-enable */
+
   & .panel-title { /* stylelint-disable-line selector-class-pattern */
     display: flex;
     border: none;
@@ -32,7 +37,6 @@
 
     .walkthrough-counter { /* stylelint-disable-line selector-class-pattern */
       flex-basis: 50%;
-      margin-top: 16px;
       text-align: right;
       color: $pf-color-black-800;
       font-size: var(--pf-global--FontSize--sm);
